Added code for custom file prefetching
This file prefetching algorithm works on the basis of access patterns of a given
file type and the cache pressure. This patch adds the code for measuring the
cache pressure, identify filetype and get the access pattern of that file.
This patch also sets up a sysctl variable to change the prefetching algorihtm.